requirements.txt
requests>=2.28.0
pandas>=1.5.0
numpy>=1.23.0
python-dotenv>=1.0.0
schedule>=1.2.0
websockets>=11.0
pykis>=1.0.0
